Fix warning, update vcxproj to v145 with portable paths, ship new exe
- Changed lastpress to ULONGLONG to match GetTickCount64() return type
- Updated all four build configurations in MouseSensChanger.vcxproj to
  use v145 toolset and $(SolutionDir)-relative include/library paths
  so the project builds on any machine without hardcoded user paths
- Added explicit interception.lib to AdditionalDependencies
- Updated Application/SensitivityRandomizer.exe (x64 Release build)

+++ Source/SensitivityRandomizer/SensitivityRandomizer.cpp
@@ -388,8 +388,8 @@ int main()
     auto   prog_start = Clock::now();
     double carryX = 0, carryY = 0;
     double sens_mult = SENS_MEAN;
-    bool   paused    = false;
-    DWORD  lastpress = 0;
+    bool      paused    = false;
+    ULONGLONG lastpress = 0;
     bool   running   = true;
 
     // Smooth mode tracking
